Always install with version 4. before it
Sadly we need this because originally this plug-in was following Cura's version numbers. With the advent of Cura's plug-in browser this is no longer necessary, but we can't reduce the version number because the plug-in will no longer show up as an update then. So we must always stick 4. in front of the version number in order to beat the highest release that followed Cura's version number: 3.2.