Support both SDK 5 and SDK 6
Cura's Marketplace only allows one version of a plug-in to be online at a time. This forces all plug-in developers to duplicate their code within the plug-in if it gets broken by the Cura API. Luckily, mine isn't broken yet so I only need to specify here that I support both SDK versions.